Move BF plot and Rowe stats into psf_diagnostics
